% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GPtail.R
\name{GPtail}
\alias{GPtail}
\title{

  Convolution of a spline density and a GP distribution

}
\usage{
GPtail(x,
       threshold.y = NA,
       distname.y = "GPD",
       par.y = c(scale = 1, shape = 0.05),
       covpar.y = NULL,
       lambda = ifelse(is.na(threshold.y), 705.8, NA),
       shift.y = ifelse(is.na(threshold.y), 0, threshold.y),
       pct.conf = c(95, 70),
       use.covlambda = "lambda" \%in\% colnames(covpar.y),
       deriv = !is.null(covpar.y),
       Tlim = c(1, 1e+05),
       pred.period = NULL,
       pred.prob = NULL,
       trace = 1L,
       N = 2 * 1024,
       N.ex = 300,
       plot.which = 0)
}
\arguments{
  \item{x}{

    An object with S3 class \code{"SplineDensity"}.

  }
  \item{threshold.y}{

    Optional threshold.

  }

  \item{distname.y}{

    Name of the distribution in the Generalized Pareto family. For now,
    only \code{"GPD"} is allowed.

  }
  \item{par.y}{

    Numeric \emph{named} vector of GP parameters, see examples. This is
    usually a vector of estimated parameters.

  }
  \item{covpar.y}{
    
    Covariance matrix to be used in a \emph{delta method}. This is for
      the parameters of the \eqn{Y}-part, but can contain a row and
      column for the event rate \code{lambda} (see Details). The
      colnames and rownames must agree and must be equal either to
      \code{names(par.y)} or to \code{c("lambda", names(par.y))}.

  }
  \item{lambda}{

    Rate to be used in the computation of the return levels. Should be
    given in \emph{events by year} since the return levels are given on
    a yearly basis.

  }
  \item{shift.y}{
 
    Shift of the distribution of \eqn{Y}, i.e. location parameter for
    the GPD.
    
  }
  \item{pct.conf}{

    Confidence levels in percent. Should be given in decreasing order.

  }
  \item{use.covlambda}{

    Logical indicating if the uncertainty on the event rate
    \code{lambda} should be taken into account in the delta method or
    not.

  }
  \item{deriv}{

    Logical. If \code{TRUE}, the derivatives of the survival of the sum
    \eqn{Z} will be computed.

  }
  \item{Tlim}{
    
    Limits for the return periods used in computations.  Only the second
    element \code{Tlim[2]} is used at present to se the maximal return
    period.

  }
  \item{pred.period}{

    If not \code{NULL}, a vector giving periods at which predictions
    (return levels and confidence limits) should be computed and returned in
    the \code{ret.lev} data.frame.
    
  }
  \item{pred.prob}{

    Alternative to \code{pred.period} to set the return periods at which
    return levels and confidence bounds are computed.
    
  }
  \item{trace}{

    Integer level of verbosity.

  }
  \item{N}{

    Number of discretization points in the range of \code{x}, hence
    length of the vector used for \code{x} in the discrete convolution.
    
  }
  \item{N.ex}{

    Number of (tail) values \eqn{z} for the r.v. q\eqn{Z}
    where an exact computation of the survival \eqn{S_Z(z)} is
    performed. The values are greater than the upper end-point of
    \eqn{X}. These values are used in most results (density,
    conditional expectation, ...) but not in the return levels
    table where only rounded return periods are used after
    an interpolation.
  }
  \item{plot.which}{
    
    Integer choosing a plot, if any. The value \code{0} plots nothing,
    \code{1} plot densities and \code{2} plot the survival..
    
  }
  \item{prob}{

    Probability for which the return levels are wanted in
    the \code{ret.lev} table. A NULL value correspond to a default
    vector of values.
  }
}

\value{
  
  A list similar to that returned by \code{\link{convSL}}.

}

\description{

  Convolution of a spline density and a GP distribution.

}

\details{

  Given a density for a bounded r.v. \eqn{X}, the distribution of
  \eqn{Z := X + Y} is computed where \eqn{Y} is a r.v. independent
  of \eqn{X} with a Generalized Pareto distribution. The
  r.v. \eqn{Y} will often be the excess in a POT model, in which
  case the exceedance rate \eqn{\lambda} should be used to get return
  periods in years.
}

\examples{
data(Brest.tide)
SD <- SplineDensity(x = Brest.tide$x, f = Brest.tide$y,
                    order = 2, nKnots = 24)
par.y <- c("scale" = rgamma(1, shape = 2, scale = 30),
           "shape" = 0.2 * runif(1))
res <- GPtail(x = SD, par.y = par.y, lambda = 1)

## plot the densities
plot(res)

## plot the return level
plot(res, which = 3)

}

\author{
  
Yves Deville

}

