% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GPtailMethods.R
\name{plot.GPtail}
\alias{plot.GPtail}
\title{Plot method for the \code{"GPtail"} class}
\usage{
\method{plot}{GPtail}(x, y = NULL, which = 1, ...)
}
\arguments{
\item{x}{An object of the \code{GPtail} class.}

\item{y}{Not used yet.}

\item{which}{

  Integer for the choice of the plot type: \code{1} plots the
  probability densities of the three r.vs \eqn{X}, \eqn{Y} and \eqn{Z},
  \code{2} plots the cumulative distribution function of \eqn{Z} and
  \code{3} shows the return level plot as produced by
  \code{\link{RSLplot}.}
  
}

\item{...}{

  Further arguments to be passed to \code{plot}. Not used
  for now.

}
}
\value{
Nothing.
}
\description{
Plot method for the \code{"GPtail"} class.
}

