% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SDMethods.R
\name{plot.SplineDensity}
\alias{plot.SplineDensity}
\title{

  Plot method for \code{SplineDensity} objects

}
\usage{
\method{plot}{SplineDensity}(x, y = NULL, ...)
}
\arguments{
  \item{x}{

    A \code{SplineDensity} object.
    
  }
  \item{y}{

    Not used yet.

  }
  \item{...}{
    
    Not used yet.
    
  }
}
\value{

  None.
  
}
\description{

  Plot method for \code{SplineDensity} objects.

}
\details{

  The density value is computed on a fine grid using the \code{predict}
  method for the class. The original points used in the creation of the
  object are shown as well.

}

