% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SDMethods.R
\name{predict.SplineDensity}
\alias{predict.SplineDensity}
\title{

  Predict (or evaluate) a Spline Density.

}
\usage{
\method{predict}{SplineDensity}(object, newdata = NULL, deriv = 0, ...)
}
\arguments{
  
  \item{object}{
    
    A \code{SplineDensity} object.

  }
  \item{newdata}{

    Numeric vector giving the values where the density (or its
    derivative) will be evaluated.

  }  
  \item{deriv}{

    Integer giving the a derivation order. It can be set to \eqn{-1} to
    get the indefinite integral.

  }
  \item{...}{

    Not used yet.
    
  }
}
\value{
  
  A list containing two numeric vectors \code{x} and \code{y}
  representing the prediction abscissae and the corresponding predicted
  values.
  
}
\description{

  Predict (or evaluate) a Spline Density.

}
\details{

  The density or one of its derivatives or the cumulative distribution
  function is evaluated at the points given in \code{newdata}.

}
\examples{
set.seed(1234)
SD <- rSplineDensity(order = 4, xmax = 10)
xg <- seq(from = -1, to = 11, length.out = 200)
p <- predict(SD, newdata = xg)
plot(p, type = "l", main = "density")
pm1 <- predict(SD, newdata = xg, deriv = -1)
plot(pm1, type = "l", main = "cdf")
}

