% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rSD.R
\name{rSplineDensity}
\alias{rSplineDensity}
\title{

  Random Spline Density

}
\usage{
rSplineDensity(knots,
               nx = 200L,
               nKnots = 10L,
               xmin = 0.0, xmax = 1.0,
               order = 4L, plot = TRUE)
}
\arguments{
  
  \item{knots}{

    Numeric vector of knots, given in increasing order.
    
  }
  \item{nx}{

    Number of default evaluation points.

  }
  \item{nKnots}{

    Number of knots if \code{knots} is not given.

  }
  \item{xmin}{

    Lower end-point of the density.

  }
  \item{xmax}{

    Upper end-point of the density.
    
  }
  \item{order}{

    Order of the spline, i.e. the degree of the polynomial pieces minus
    one. With \code{ord = 2} one gets a broken line spline and
    \code{ord = 4} leads to a cubic spline.

  }
  \item{plot}{

    Logical. If \code{TRUE} a plot is shown.

  }
}
\value{

  An object of class \code{"SplineDensity"}.

}
\description{

  Random Spline Density.

}
\details{

  From the sequence of knots and the order, a basis of B-splines is
  built. A vector of positive random coefficients is drawn and suitably
  normalized to ensure that the spline integrates to one on the support.
  
}
\examples{
set.seed(1234)
SD <- rSplineDensity(order = 4, xmax = 10)
plot(SD, main = "spline density")
## uniform knots sequence
SDu <- rSplineDensity(order = 4, knots = 0:10)
plot(SDu, main = "spline density with uniform knots sequence.")
}
\author{
Yves Deville
}

