\name{SeaLev-package}
\Rdversion{1.1}
\alias{SeaLev-package}
\alias{SeaLev}
\docType{package}
\title{
 
  Convolution Method for Sea Levels

}
\encoding{UTF-8}
\description{
  
  Derive the probability distribution for the sea level from a
  distribution for the \dfn{tidal component} and a distribution for the
  \dfn{surge component}.
  
  The package development was initiated, directed and financed by the
  french \emph{Institut de radioprotection et de \enc{sûreté}{surete}
  \enc{nucléaire}{nucleaire}} (IRSN).

}
\details{

  The DESCRIPTION file:
  \packageDESCRIPTION{SeaLev}

  %% \packageIndices{SeaLev}
  
}
\author{
  
  \packageAuthor{SeaLev}

  Maintainer: \packageMaintainer{SeaLev}

}
\references{

  The convolution method or \emph{Joint Probability Method} (JPM) was
  initiated by Pugh and Vassie in

  D. T. Pugh and J. M. Vassie
  "Extreme sea-levels from tide and surge probability" \emph{Proceeding
  of the 16-th Coastal Engineering Conference}, 1978, Hamburg.
  
  The method is described and illustrated in the book (in french) by
  B. Simon

  B. Simon. \emph{La \enc{marée}{maree} \enc{océanique}{oceanique}
  \enc{côtière}{cotiere}}.  Institut
  \enc{océanographique}{oceanographique}, 2007.
  
}
\keyword{package}

\seealso{
  
  The \code{\link[Renext:Renext-package]{Renext}} package for POT
  fitting of various distributions.

}
%%\examples{}
