\name{varnames}
\alias{varnames}

\title{

  S3 generic function to extract the names of variables
  
}

\description{

  Extracts the names of variables from an object.

}

\usage{

varnames(object, ...)

}
\arguments{
  \item{object}{
    
    The object from which to extract variables names. 
    
  }
  \item{\dots}{
    
    Other arguments that can be used for some methods.
    
  }
}
%% \details{ }
\value{

  A character vector representing the names or descriptions
  of the variables.
  
}
%% \references{}
%% \author{}
%% \note{}

\seealso{

  \code{\link{varnames<-}}

}
\keyword{models}
